Updated main.rs
Removed commented out old code.
Added commenting
Changed num_threads from u128 to u16.

+++ src/main.rs
@@ -10,15 +10,30 @@ use std::str::FromStr;
 use num_integer::Integer;
 
 fn main() {
+    // Setting the start number for the program, Must be Odd!
     let start: BigUint = BigUint::from_str("1300000000000000001").unwrap_or_else(|_| Zero::zero());
+    // Setting the end number for the program
     let end: BigUint = BigUint::from_str("1700000000000000001").unwrap_or_else(|_| Zero::zero());
+    // Wrapping the BigUints in Arc.
     let start = Arc::new(start);
     let end = Arc::new(end);
-    let num_threads = num_cpus::get() as u128;
+    // Get the number of threads on the cpu, this will be what creates the
+    let num_threads = num_cpus::get() as u16;
+    // print out the threads.
     println!("Using {num_threads} threads");
+    // begin the range
     check_range_for_primes(start,end);
 }
 
+
+/// This function just goes through the given range and starts prime_checks on the individual indices.
+///
+/// # Arguments
+///
+/// * `start`: The start of the range of numbers that are to be tested for primeness. Expects Arc<BigUint>
+/// * `end`: The end of the range of numbers that are to be tested for primeness. Expects Arc<BigUint>
+///
+/// returns: ()
 fn check_range_for_primes(start: Arc<BigUint>, end:Arc<BigUint>){
     let mut it = (*start).clone();
     println!("PRIMES:");
@@ -33,6 +48,13 @@ fn check_range_for_primes(start: Arc<BigUint>, end:Arc<BigUint>){
 }
 
 
+/// Tests if a number given is prime by utilizing division proofing.
+///
+/// # Arguments
+///
+/// * `x_arc`: Number to be tested for primality. Expects reference to BigUint
+///
+/// returns: bool
 fn prime_check(x_arc: &BigUint) -> bool {
     let x = &*x_arc;
 
@@ -48,7 +70,6 @@ fn prime_check(x_arc: &BigUint) -> bool {
     let segment_size = (&sqrt_x / num_threads) + 1u8;
     let mut handles = vec![];
     let found_divisor = Arc::new(AtomicBool::new(false));
-    //let x = Arc::new(x);
 
     for i in 0..num_threads {
         let x_clone = x_arc.clone();
@@ -79,6 +100,14 @@ fn prime_check(x_arc: &BigUint) -> bool {
     !found_divisor.load(Ordering::Relaxed)
 }
 
+
+/// Finds the approximate square root of a given number.
+///
+/// # Arguments
+///
+/// * `x`: Number of which to find the square root of. Expects reference to BigUint.
+///
+/// returns: BigUint
 fn approximate_sqrt(x:&BigUint) -> BigUint {
     let two = BigUint::from(2u8);
     let mut s = x.clone();
@@ -90,4 +119,3 @@ fn approximate_sqrt(x:&BigUint) -> BigUint {
     }
     s
 }
-
